+++ lib/src/collect.dart
@@ -31,9 +31,12 @@ const _retryInterval = Duration(milliseconds: 200);
 ///
 /// If [scopedOutput] is non-empty, coverage will be restricted so that only
 /// scripts that start with any of the provided paths are considered.
+///
+/// if [isolateIds] is set, the coverage gathering will be restricted to only
+/// those VM isolates.
 Future<Map<String, dynamic>> collect(Uri serviceUri, bool resume,
     bool waitPaused, bool includeDart, Set<String> scopedOutput,
-    {Duration timeout}) async {
+    {Set<String> isolateIds, Duration timeout}) async {
   scopedOutput ??= Set<String>();
   if (serviceUri == null) throw ArgumentError('serviceUri must not be null');
 
@@ -63,7 +66,8 @@ Future<Map<String, dynamic>> collect(Uri serviceUri, bool resume,
       await _waitIsolatesPaused(service, timeout: timeout);
     }
 
-    return await _getAllCoverage(service, includeDart, scopedOutput);
+    return await _getAllCoverage(
+        service, includeDart, scopedOutput, isolateIds);
   } finally {
     if (resume) {
       await _resumeIsolates(service);
@@ -72,13 +76,14 @@ Future<Map<String, dynamic>> collect(Uri serviceUri, bool resume,
   }
 }
 
-Future<Map<String, dynamic>> _getAllCoverage(
-    VmService service, bool includeDart, Set<String> scopedOutput) async {
+Future<Map<String, dynamic>> _getAllCoverage(VmService service,
+    bool includeDart, Set<String> scopedOutput, Set<String> isolateIds) async {
   scopedOutput ??= Set<String>();
   final vm = await service.getVM();
   final allCoverage = <Map<String, dynamic>>[];
 
   for (var isolateRef in vm.isolates) {
+    if (isolateIds != null && !isolateIds.contains(isolateRef.id)) continue;
     if (scopedOutput.isNotEmpty) {
       final scripts = await service.getScripts(isolateRef.id);
       for (var script in scripts.scripts) {